[FIX] sale: fix quantity non visible in portal
Create a sale order
Put an optional product with long description (no line break)
Save and confirm
Log in the portal with the receiving user, add the optional product to
the order.

The customer can't see the quantity of the optional product.
The number is not showing in the quantity adjustment widet.

opw-2304224

closes odoo/odoo#58149

X-original-commit: 1b4af52d3200c26646aa383ef34976cb8b16fa71
Signed-off-by: agr-odoo <agr-odoo@users.noreply.github.com>